+++ S4.c
@@ -2,7 +2,7 @@
 
 int main(){
     int max;
-    float sum = 0;
+    double sum = 0;
     scanf("%d",&max);
     for(int i=1;i<=max;i+2){
         sum += 1.0/(i*(i+2));
