+++ S6.c
@@ -2,7 +2,7 @@
 
 int main(){
     int max;
-    float sum = 1;
+    double sum = 1;
     scanf("%d",&max);
     for(int i=2;i<=max;i++){
         sum += 1.0/(i*i*i*i);
